add plan issue issueThread ride

+++ src/api/routes/Plan.ts
@@ -1,32 +1,30 @@
 import { Router } from 'express';
 import { celebrate, Joi } from 'celebrate';
-import { attachUser, checkRole, isAuth } from '../middlewares';
+import { checkRole, isAuth } from '../middlewares';
 import { Role } from '../entities/User';
-import ArticleService from '../services/ArticleService';
 import { Container } from 'typedi';
-import { userRequest } from '../../types/userRequest';
-import { Article } from '../entities/Article';
+import PlanService from '../services/PlanService';
+import SubscriptionService from '../services/SubscriptionService';
 
 const route = Router();
 const paramsRules = celebrate({
   body: Joi.object({
-    title: Joi.string().max(32).min(10).required(),
-    content: Joi.string().required(),
-    cover: Joi.string().required(),
+    name: Joi.string().max(32).min(10).required(),
+    price: Joi.number().required(),
+    costPerMinute: Joi.number().required(),
+    isUnlimited: Joi.boolean().required(),
   }),
 });
-const defaultService = ArticleService;
-const basePath = '/article/';
+const defaultService = PlanService;
+const basePath = '/plan/';
 
 route.post(
   basePath,
   isAuth,
-  checkRole(Role.STAFF),
-  attachUser,
+  checkRole(Role.ADMIN),
   paramsRules,
-  async (req: userRequest, res, next) => {
+  async (req, res, next) => {
     const service = Container.get(defaultService);
-    req.body.author = req.currentUser.id;
     try {
       const entityResult = await service.create(req.body);
       return res.status(201).json(entityResult);
@@ -41,14 +39,8 @@ route.get(basePath, async (req, res, next) => {
     const service = Container.get(defaultService);
     const offset = req.body.offset || 0;
     const limit = req.body.limit || 20;
-    let articles: Article[];
-    if (typeof req.body.tags !== 'undefined') {
-      const tags = escape(req.body.tags);
-      articles = await service.searchByName(tags, { offset, limit });
-    } else {
-      articles = await service.find({ offset, limit });
-    }
-    return res.status(200).json(articles);
+    const result = await service.find({ offset, limit });
+    return res.status(200).json(result);
   } catch (e) {
     return next(e);
   }
@@ -68,11 +60,22 @@ route.get(basePath + ':id', async (req, res, next) => {
 route.delete(
   basePath + ':id',
   isAuth,
-  checkRole(Role.STAFF),
+  checkRole(Role.ADMIN),
   async (req, res, next) => {
     try {
       const service = Container.get(defaultService);
+      const dependencyService = Container.get(SubscriptionService);
       const id = Number.parseInt(req.params.id);
+      const dependency = await dependencyService.getAllFromPlan(id, {
+        limit: 1,
+        offset: 0,
+      });
+      if (dependency.length > 0) {
+        res
+          .status(403)
+          .json({ message: 'Impossible de supprimmer ce forfait' });
+        return;
+      }
       await service.delete(id);
       return res.status(204);
     } catch (e) {
@@ -84,7 +87,7 @@ route.delete(
 route.put(
   basePath + ':id',
   isAuth,
-  checkRole(Role.STAFF),
+  checkRole(Role.ADMIN),
   paramsRules,
   async (req, res, next) => {
     const service = Container.get(defaultService);
